+++ src/twg2/dependency/eclipseProject/EclipseClasspathDoc.java
@@ -6,6 +6,7 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.UncheckedIOException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -19,15 +20,20 @@ import javax.xml.transform.stream.StreamResult;
 import lombok.val;
 
 import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
+
+import twg2.collections.dataStructures.SortedList;
+import twg2.text.stringUtils.StringCheck;
 
 /**
  * @author TeamworkGuy2
  * @since 2016-2-3
  */
 public class EclipseClasspathDoc implements EclipseClasspathEntries {
-	private static final DocumentBuilderFactory docBldrFactory = DocumentBuilderFactory.newInstance();;
+	private static final DocumentBuilderFactory docBuilderFactory = DocumentBuilderFactory.newInstance();
 
 	File file;
 	Document doc;
@@ -38,6 +44,7 @@ public class EclipseClasspathDoc implements EclipseClasspathEntries {
 		this.file = file;
 		this.doc = doc;
 		this.classPathEntries = classPathEntries;
+		Collections.sort(this.classPathEntries);
 	}
 
 
@@ -63,6 +70,27 @@ public class EclipseClasspathDoc implements EclipseClasspathEntries {
 	}
 
 
+	public void addClassPathEntry(ClassPathEntry entry) {
+		int insertPoint = SortedList.addItem(this.classPathEntries, entry);
+		Node cps = this.doc.getElementsByTagName("classpath").item(0);
+
+		Element entryNode = this.doc.createElement("classpathentry");
+		entryNode.setAttribute("kind", entry.getKind());
+		entryNode.setAttribute("path", entry.getPath());
+
+		if(!StringCheck.isNullOrEmpty(entry.getSourcePath())) { entryNode.setAttribute("sourcepath", entry.getSourcePath()); }
+		val siblings = NodeUtil.filter(cps.getChildNodes(), (n) -> n.getNodeType() == Node.ELEMENT_NODE);
+		val indentationText = this.doc.createTextNode(ClassPathEntry.inbetweenElementText);
+		NodeUtil.insertAfter(cps, siblings, insertPoint, entryNode);
+		NodeUtil.insertAfter(cps, siblings, insertPoint + 1, indentationText);
+	}
+
+
+	public void removeClassPathEntry(ClassPathEntry entry) {
+		this.classPathEntries.remove(entry);
+	}
+
+
 	public void saveXmlDom(OutputStream out) throws TransformerException {
 		val writeFactory = TransformerFactory.newInstance();
 		val writer = writeFactory.newTransformer();
@@ -75,7 +103,7 @@ public class EclipseClasspathDoc implements EclipseClasspathEntries {
 
 	public static final EclipseClasspathDoc loadXmlDom(File srcFile, InputStream in) {
 		try {
-			val docBldr = docBldrFactory.newDocumentBuilder();
+			val docBldr = docBuilderFactory.newDocumentBuilder();
 			val doc = docBldr.parse(in);
 			NodeList cps = doc.getElementsByTagName("classpathentry");
 			List<ClassPathEntry> cpEntries = new ArrayList<>();